Fix bug in calibration that results in empty Sv (#755)
* add first stage fix for Environment.time1 vs Beam_groupX.ping_time discrepancy

* make sure changes work with ek60 data in echodata.compute_range

* factor out harmonize time function to just in echodata

* add harmonize time to azfp cal function

* remove unused function, add typing

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update echopype/echodata/echodata.py


* move EnvParams to calibrate/env_params.py

* actually adding calibrate/env_params.py

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix logic in harmonize time function

* remove sonar_model check and just interpolate, add test

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix calibrate/env_params.py so all env_params have coordinate time1 instead of ping_time

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix test name and param: ek60 always only has power samples

* add back checking ek60/ek80 case in compute_range

* drop time1 for calibrate/compute_range ek60 outputs since it is remnant of interpolation

* fix test_env_params ping_time-->time1

* revise docstring
