+++ echopype/tests/echodata/test_echodata.py
@@ -6,7 +6,7 @@ from datatree import DataTree
 from zarr.errors import GroupNotFoundError
 
 import echopype
-from echopype.calibrate.calibrate_base import EnvParams
+from echopype.calibrate.env_params import EnvParams
 from echopype.echodata import EchoData
 from echopype import open_converted
 
@@ -68,7 +68,7 @@ def ek60_converted_zarr(request, test_path):
             None,
             None,
             "CW",
-            "complex",
+            "power",
         ),
         (
             ("EK80_NEW", "D20211004-T233354.raw"),
@@ -128,7 +128,7 @@ def ek60_converted_zarr(request, test_path):
         ),
     ],
     ids=[
-        "ek60_cw_complex",
+        "ek60_cw_power",
         "ek80_cw_power",
         "ek80_bb_complex",
         "ek80_cw_complex",
